requirements.txt
# ============================================================================
# FEAnalyzer — Production Python dependencies
# ============================================================================
# Installed into the Docker image once and shared by the `web` and `celery`
# services. Versions pinned to known-stable releases (late 2025).
# ============================================================================

# --- Django core ---------------------------------------------------------
Django==5.1.4
django-environ==0.11.2
django-cors-headers==4.6.0
whitenoise==6.8.2

# --- REST API ------------------------------------------------------------
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# --- Async tasks ---------------------------------------------------------
celery==5.4.0
redis==5.2.1

# --- Database driver -----------------------------------------------------
psycopg[binary]==3.2.3

# --- WSGI server ---------------------------------------------------------
gunicorn==23.0.0

# --- Scraping / analysis (existing logic) --------------------------------
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# --- API documentation (OpenAPI 3 / Swagger UI) --------------------------
# [sidecar] bundles Swagger UI & ReDoc assets so the /api/docs/ page works
# without a CDN — important for offline demo / air-gapped VPS deployments.
drf-spectacular[sidecar]==0.27.2

# --- Report export --------------------------------------------------------
openpyxl==3.1.5
# PDF generation for /api/reports/<id>/export.pdf/ — pure-Python, no native deps,
# so the Docker image build stays fast and the wheel is platform-independent.
reportlab==4.2.5
